Sort Array of very large size(Merge/Insertion)Sort
MergeSort continuously divides the data into segments until segments of size 1 are reached. It then begins the merging phase.So thought improvements could be made if we could reduce the cost of merging.So I came with idea to combine MergeSort and InsertionSort to reduce the over head of merging. To do this we terminate the MergeSort division when segments of some given size are reached, use InsertionSort to sort the segments and then do the merging as before